Split apart grammar construction from ast construction and validation.
Proof of concept exploring a different direction:

Most of the previous patches were predicated on a malfored `GrammarAST`
never escaping `YaccGrammar::new_with_storaget`. This resulted in an
API where most `analyze_*` calls could be called multiple times with
different analysis passes except for `analyze_ast`, which could only
be called once.

Further, it put `Analysis` in an awkward position of being called
from a places where there are different return values unrelated to
analysis. Relying on the owner to call back into the analysis afterwords
to obtain the results.

While much of the current API still very much reflects these constraints.
This patch explores a new direction: return an AST marking it with
it's validity which is either invalid or valid, then pass it back in to
create a `YaccGrammar`. This then allows calls analyze to be moved
entirely into `CTBuilder`, and could allow `analyze` to be called
multiple times with an `AST`, and return things.

This patch is still a bit awkward though:
    Since errors get returned from `ast_validation`, passed into
    `new_with_storaget` merely so that they can be returned again.

    but it is a start to play around with other API's...